Backed out 3 changesets (bug 1459785) for causing https://bugzilla.mozilla.org/show_bug.cgi?id=1464089 a=backout
Backed out changeset 88675b68241a (bug 1459785)
Backed out changeset b7c91a6f1b0a (bug 1459785)
Backed out changeset 21af8dc00aa8 (bug 1459785)

+++ gfx/angle/checkout/src/libANGLE/renderer/d3d/BufferD3D.h
@@ -55,7 +55,7 @@ class BufferD3D : public BufferImpl
     virtual void initializeStaticData(const gl::Context *context);
     virtual void invalidateStaticData(const gl::Context *context);
 
-    void promoteStaticUsage(const gl::Context *context, size_t dataSize);
+    void promoteStaticUsage(const gl::Context *context, int dataSize);
 
     gl::Error getIndexRange(const gl::Context *context,
                             GLenum type,
@@ -80,7 +80,7 @@ class BufferD3D : public BufferImpl
     StaticIndexBufferInterface *mStaticIndexBuffer;
     unsigned int mStaticBufferCacheTotalSize;
     unsigned int mStaticVertexBufferOutOfDate;
-    size_t mUnmodifiedDataUse;
+    unsigned int mUnmodifiedDataUse;
     D3DBufferUsage mUsage;
 };
